added readme and jwt implementation completed.

+++ Controllers/FollowController.js
@@ -2,17 +2,35 @@ const db = require('../utils/db');
 
 const newFollow = async (req,res) => {
     try{
-        const {user_id_follow,user_id_following} = req.body;
+        const {user_id_following} = req.body;
+        const user_id_follow = req.body.user.id;
+        if(user_id_follow===user_id_following){
+            return res.status(404).json({
+                status: "Can't follow yourself"
+            });
+        }
+        const count1 = await db.User.count({where:{id:user_id_following}});
+        if(count1===0){
+            return res.status(404).json({
+                status: "user does not exist",
+            });
+        }
+        const count = await db.Follows.count({where:{user_id_follow,user_id_following}});
+        if(count!==0){
+            return res.status(404).json({
+                status: "follow already exists",
+            });
+        }
         const newfollow = await db.Follows.create({
             user_id_follow,user_id_following
         });
         console.log(newfollow);
-        res.status(200).json({
+        return res.status(200).json({
             status: 'success, new follow created'
         });
     }catch(err){
         console.log(err);
-        res.status(404).json({
+        return res.status(404).json({
             status: "newfollow catch error",
         });
     }
@@ -20,14 +38,27 @@ const newFollow = async (req,res) => {
 
 const unFollow = async (req,res) => {
     try{
-        const {user_id_follow,user_id_following} = req.body;
+        const {user_id_following} = req.body;
+        const user_id_follow = req.body.user.id;
+        const count1 = await db.User.count({where:{id:user_id_following}});
+        if(count1===0){
+            return res.status(404).json({
+                status: "user does not exist",
+            });
+        }
+        const count = await db.Follows.count({where:{user_id_follow,user_id_following}});
+        if(count===0){
+            return res.status(404).json({
+                status: "follow does not exists",
+            });
+        }
         await db.Follows.destroy({where:{user_id_follow,user_id_following}});
-        res.status(200).json({
+        return res.status(200).json({
             status: 'success, follow deleted'
         });
     }catch(err){
         console.log(err);
-        res.status(404).json({
+        return res.status(404).json({
             status: "unfollow catch error",
         });
     }
@@ -35,7 +66,7 @@ const unFollow = async (req,res) => {
 
 const getFollows = async (req,res) => {
     try{
-        const {user_id_follow} = req.body;
+        const user_id_follow = req.body.user.id;
         const follows = await db.Follows.findAll({
             where: {user_id_follow},
             order: [['createdAt','DESC']]
@@ -54,7 +85,7 @@ const getFollows = async (req,res) => {
 
 const getFollowings = async (req,res) => {
     try{
-        const {user_id_following} = req.body;
+        const user_id_following = req.body.user.id;
         const followings = await db.Follows.findAll({
             where: {user_id_following},
             order: [['createdAt','DESC']]
